+++ web/app/themes/lark-child/config/theme.php
@@ -19,6 +19,7 @@ return [
         'mailgun/mailgun.php',
         'contact-form-7/wp-contact-form-7.php',
         'wp-pagenavi/wp-pagenavi.php',
+        'tinymce-advanced/tinymce-advanced.php',
 
         // Used for duplicating pages as well as the ACF on the page,
         // very useful for when copying the homepage on dev
@@ -39,14 +40,32 @@ return [
 
     'fonts' => [
         'h1' => [
-            'size' => '',
-            'font' => '',
-            'margin' => '',
+            'size' => 'text-32 md:text-42',
+            'font' => 'font-questrial',
+            'margin' => 'mb-30',
         ],
 
         'h2' => [
-            'size' => '',
-            'font' => '',
+            'size' => 'text-26 leading-36 -tracking-0.6 md:text-36 md:-tracking-0.4 md:leading-45 md:-word-spacing-2.6',
+            'font' => 'font-questrial',
+            'margin' => 'mb-30 md:mb-40',
+        ],
+
+        'h3' => [
+            'size' => 'text-24 leading-32',
+            'font' => 'font-questrial',
+            'margin' => 'mb-30',
+        ],
+
+        'h4' => [
+            'size' => 'text-18',
+            'font' => 'font-questrial',
+            'margin' => '',
+        ],
+
+        'h5' => [
+            'size' => 'text-16',
+            'font' => 'font-questrial',
             'margin' => '',
         ],
     ],
@@ -65,5 +84,22 @@ return [
     */
 
     'h' => 1,
+
+    /*
+    |--------------------------------------------------------------------------
+    | TinyMCE editor
+    |--------------------------------------------------------------------------
+    */
+
+    'tinymce' => [
+        'hide-unused-buttons' => [
+            'fontselect',
+            'fontsizeselect',
+            'blockquote',
+            'wp_more',
+            'textcolor',
+            'colorpicker',
+        ],
+    ],
 
 ];